rename stores file and move to stores folder
