+++ packages/studio-components/src/components/editor-split-view/editor-split-view.tsx
@@ -1,6 +1,6 @@
 import { Component, Host, h } from '@stencil/core';
 import Split from 'split.js'
-import state from "../../utils/store";
+import state from "../../stores/editor_store";
 
 
 @Component({
